# src/backend/cube/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  ModelInfoResponse:
    properties:
      id:
        example: f6457bdf-4e67-4f05-9108-1cbc0fec9405
        type: string
      name:
        example: awesome_username
        type: string
      structure: {}
    type: object
  ModelWeightsInfoResponse:
    properties:
      id:
        example: f6457bdf-4e67-4f05-9108-1cbc0fec9405
        type: string
      name:
        example: awesome_username
        type: string
      offsets: {}
      weights: {}
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is cube server.
  license:
    name: MIT
    url: https://mit-license.org/
  title: Cube API
  version: "1.0"
paths:
...
  /api/v1/models:
    delete:
      description: Deletes model info owned by authorized user
      parameters:
      - description: Model ID to delete
        in: query
        name: id
        type: string
      responses:
        "200":
          description: Model info deleted
        "400":
          description: Invalid request
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to delete model info from storage
      summary: Delete model info
      tags:
      - user
    get:
      description: Find such model info as id, username, email and fullname
      parameters:
      - description: Model ID to search for
        in: query
        name: id
        type: string
      - description: User ID that owns model to search for
        in: query
        name: owner_id
        type: string
      - description: Model name to search for
        in: query
        name: name
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Page size for pagination
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: Model info found
          schema:
            items:
              $ref: '#/definitions/ModelInfoResponse'
            type: array
        "400":
          description: Invalid request
        "500":
          description: Failed to get model info from storage
      summary: Find model info
      tags:
      - user
    patch:
      consumes:
      - multipart/form-data
      description: Update such model info as weights, weights titles
      parameters:
      - description: Model ID to update
        in: formData
        name: id
        required: true
        type: string
      - description: Model Weights ID to update
        in: formData
        name: weights_id
        type: string
      - description: Model Weights Title to set
        in: formData
        name: weights_title
        type: string
      - description: Model Weights to Update/Add
        in: formData
        name: weights
        type: file
      responses:
        "200":
          description: Model info updated
        "400":
          description: Invalid request
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to update model info
      summary: Update model info
      tags:
      - user
    post:
      consumes:
      - multipart/form-data
      description: Adds such model info as title, structure, weights
      parameters:
      - description: Model Title to create
        in: formData
        name: title
        required: true
        type: string
      - description: Model Structure Title to add
        in: formData
        name: structure_title
        required: true
        type: string
      - description: Model Structure to add
        in: formData
        name: structure
        required: true
        type: file
      - description: Model Weights Title to add
        in: formData
        name: weights_title
        required: true
        type: string
      - description: Model Weights to add
        in: formData
        name: weights
        required: true
        type: file
      responses:
        "200":
          description: Model created
        "400":
          description: Invalid request
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to create model info
      summary: Create new model
      tags:
      - user
  /api/v1/models/weights:
    delete:
      description: Deletes model info from any user
      parameters:
      - description: Model ID to delete
        in: query
        name: id
        type: string
      responses:
        "200":
          description: Model info deleted
        "400":
          description: Invalid request
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to delete model info from storage
      summary: Delete model info
      tags:
      - user
    get:
      description: Find such model info as id, username, email and fullname
      parameters:
      - description: Weight ID to search for
        in: query
        name: id
        type: string
      - description: Structure ID to search for
        in: query
        name: structure_id
        type: string
      - description: Weights name to search for
        in: query
        name: name
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: integer
      - description: Page size for pagination
        in: query
        name: per_page
        type: integer
      responses:
        "200":
          description: Model weights info found
          schema:
            items:
              $ref: '#/definitions/ModelWeightsInfoResponse'
            type: array
        "400":
          description: Invalid request
        "500":
          description: Failed to get model weights info from storage
      summary: Find model info
      tags:
      - admin
    patch:
      consumes:
      - multipart/form-data
      description: Update such model info as weights, weights titles
      parameters:
      - description: Model Weights ID to update
        in: formData
        name: weights_id
        type: string
      - description: Model Weights Title to set
        in: formData
        name: weights_title
        type: string
      - description: Model Weights to Update/Add
        in: formData
        name: weights
        type: file
      responses:
        "200":
          description: Model info updated
        "400":
          description: Invalid request
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to update model info
      summary: Update model info
      tags:
      - user
    post:
      consumes:
      - multipart/form-data
      description: Adds model weights info to existing model
      parameters:
      - description: Model ID to add weights to
        in: formData
        name: id
        required: true
        type: string
      - description: Model Weights Title to add
        in: formData
        name: weights_title
        required: true
        type: string
      - description: Model Weights to add
        in: formData
        name: weights
        required: true
        type: file
      responses:
        "200":
          description: Weights added
        "400":
          description: Invalid request
        "403":
          description: Invalid token, user id not specified
        "500":
          description: Failed to create model weights info
      summary: Create new model weights info
      tags:
      - user
...
securityDefinitions:
  Bearer:
    type: basic
swagger: "2.0"